+++ commander/sync.action.ts
@@ -14,6 +14,7 @@ export default async function(): Promise<void> {
 	if (globalOpts.verbose) logger.enableVerboseMode();
 
 	const actionOpts = syncCmd.opts<SyncOptions>();
+	const { jiraEstimateField } = actionOpts;
 
 	logger.debug(logname, "parsed options", { actionOpts });
 
@@ -63,6 +64,7 @@ export default async function(): Promise<void> {
 				gh,
 				project,
 				item,
+				jiraEstimateField,
 			})
 		}
 
@@ -129,9 +131,15 @@ async function createJiraIssueFromGhTaskWithoutUrl(args: {
 	jira: JiraClient,
 	item: Item,
 	project: Project,
+	jiraEstimateField?: string;
 }) {
 	const { gh, jira, item, project } = args;
 	const actionOpts = syncCmd.opts<SyncOptions>();
+
+	if (!item[itemField.STATUS]) {
+		logger.warn(logname, "item does not have any status, assuming it is not ok and skipping creation", { itemId: item.id, title: item[itemField.TITLE].value })
+		return;
+	}
 
 	if (!item[itemField.JIRA_ISSUE_TYPE].value) {
 		logger.warn(logname, "item does not have an issue type, skipping creation", { itemId: item.id, title: item[itemField.TITLE].value })
@@ -147,12 +155,18 @@ async function createJiraIssueFromGhTaskWithoutUrl(args: {
 		summary = item[itemField.TITLE].value;
 	}
 
-	const [createRes, createErr] = await jira.issues.create({
+	const createOpts = {
 		projectKey: actionOpts.jiraProjectKey,
 		summary,
 		issueName: item[itemField.JIRA_ISSUE_TYPE].value,
 		accountId,
+	}
+
+	if (actionOpts.jiraEstimateField && item[itemField.ESTIMATE].value) Object.assign(createOpts, {
+		storyPointsField: { name: actionOpts.jiraEstimateField, points: item[itemField.ESTIMATE].value }
 	});
+
+	const [createRes, createErr] = await jira.issues.create(createOpts);
 
 	if (createErr) {
 		logger.error(logname, "unable to create jira issue", { itemId: item.id, err: createErr });