+++ commander/sync.cmd.ts
@@ -11,13 +11,14 @@ export type SyncOptions = {
 	transitionsToDone?: number[],
 	sleepTime?: number,
 	jiraIssuePrefix?: string,
+	jiraEstimateField?: string,
 }
 
 const syncCmdName = "sync";
 const syncCmd = new Command(syncCmdName);
 syncCmd.description("sync GitHub project tickets with Jira");
 
-const transitionsToWipOption = new Option("--transitions-to-wip <NUMBER,...>", "list of jira issue transitions in order to have a wip task");
+const transitionsToWipOption = new Option("--transitions-to-wip <number,...>", "list of jira issue transitions in order to have a wip task");
 transitionsToWipOption.argParser<SyncOptions["transitionsToWip"]>((value, _prev) => {
 	const [option, error] = util.numberArrayFromString(value, ",");
 	if (error) return util.error(error);
@@ -25,7 +26,7 @@ transitionsToWipOption.argParser<SyncOptions["transitionsToWip"]>((value, _prev)
 });
 syncCmd.addOption(transitionsToWipOption);
 
-const transitionsToDoneOption = new Option("--transitions-to-done <NUMBER,...>", "list of jira issue transitions in order to have a done task");
+const transitionsToDoneOption = new Option("--transitions-to-done <number,...>", "list of jira issue transitions in order to have a done task");
 transitionsToDoneOption.argParser<SyncOptions["transitionsToWip"]>((value, _prev) => {
 	const [option, error] = util.numberArrayFromString(value, ",");
 	if (error) return util.error(error);
@@ -33,7 +34,7 @@ transitionsToDoneOption.argParser<SyncOptions["transitionsToWip"]>((value, _prev
 });
 syncCmd.addOption(transitionsToDoneOption);
 
-const mapGhAssigneeOption = new Option("--gh-assignees-map <GH_USER:JIRA_USER,...>", "map of GitHub users to Jira ones (email)");
+const mapGhAssigneeOption = new Option("--gh-assignees-map <gh_user:jira_user,...>", "map of GitHub users to Jira ones (email)");
 mapGhAssigneeOption.argParser<SyncOptions["ghAssigneesMap"]>((value, _prev) => {
 	const map: Record<string, string | undefined> = {}
 	const commaSplitted = value.split(",");
@@ -48,14 +49,15 @@ mapGhAssigneeOption.argParser<SyncOptions["ghAssigneesMap"]>((value, _prev) => {
 });
 syncCmd.addOption(mapGhAssigneeOption);
 syncCmd.option("--sleep-time <ms>", "sleep time between executions. If not specified the program will run once");
-syncCmd.requiredOption("--gh-project-id <STRING>", "Github project ID");
-syncCmd.requiredOption("--jira-project-key <STRING>", "Jira project KEY");
-syncCmd.requiredOption("--jira-subdomain <STRING>", "Jira subdomain");
-
+syncCmd.requiredOption("--gh-project-id <string>", "Github project ID");
+syncCmd.requiredOption("--jira-project-key <string>", "Jira project KEY");
+syncCmd.requiredOption("--jira-subdomain <string>", "Jira subdomain");
 // Jira issue prefix to be appended to title
-const jiraIssuePrefixOption = new Option("--jira-issue-prefix <STRING>", "Prefix to be added to jira issue title");
-jiraIssuePrefixOption.required = false;
+const jiraIssuePrefixOption = new Option("--jira-issue-prefix <string>", "prefix to be added to jira issue title");
 syncCmd.addOption(jiraIssuePrefixOption);
+
+const jiraEstimateFieldOpt = new Option("--jira-estimate-field <string>", "Jira field name that holds estiamte value");
+syncCmd.addOption(jiraEstimateFieldOpt);
 
 syncCmd.action(syncAction);