+++ entrypoint.sh
@@ -21,6 +21,7 @@ bun run index.ts ${VERBOSE_FLAG} sync \
 	--jira-subdomain=${JIRA_SUBDOMAIN} \
 	--jira-project-key=${JIRA_PROJECT_KEY} \
 	--jira-issue-prefix=${JIRA_ISSUE_PREFIX} \
+	--jira-estiamte-field=${JIRA_ESTIMATE_FIELD} \
 	--transitions-to-wip=${JIRA_WIP_TRANSITIONS} \
 	--transitions-to-done=${JIRA_DONE_TRANSITIONS} \
 	--sleep-time=${SLEEP_TIME}