+++ CHANGELOG.md
@@ -11,11 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docker compose volume to persist executions data.
 - Setting up the `--jira-issue-prefix=<string>` pre-appends the given prefix to jira issue titles.
 - New `--user <string>` option in the list github-projects command.
+- New `--jira-estimate-field <string>` option that allow users to specify the jira issue field name that stores story points (estimate).
 
 ### Changed
 - Retrieving items from the local storage with(out) an url now checks if the url field is a valid url. This will cause that item's urls that are invalid are going to be repleaced with a valid url that will point to a new jira issue.  
 - `github-projects` command was renamed to `github-project`.
   - `listOrganization` sub command was renamed to `list`.
+- GitHub issues that lack of status no longer creates Jira issues to prevent malformed Jira issues.
+
+### Fixed
+- GitHub task's estimate field now is reflected in Jira issues.
 
 ## [v0.3.0]
 ### Added